feat(store): resize PackedCounter use 5 bytes per length

+++ packages/store/src/codegen/tables/Callbacks.sol
@@ -192,8 +192,8 @@ library Callbacks {
 
   /** Tightly pack full data using this table's schema */
   function encode(bytes24[] memory value) internal view returns (bytes memory) {
-    uint16[] memory _counters = new uint16[](1);
-    _counters[0] = uint16(value.length * 24);
+    uint40[] memory _counters = new uint40[](1);
+    _counters[0] = uint40(value.length * 24);
     PackedCounter _encodedLengths = PackedCounterLib.pack(_counters);
 
     return abi.encodePacked(_encodedLengths.unwrap(), EncodeArray.encode((value)));
