+++ packages/store/src/codegen/tables/Mixed.sol
@@ -471,9 +471,9 @@ library Mixed {
 
   /** Tightly pack full data using this table's schema */
   function encode(uint32 u32, uint128 u128, uint32[] memory a32, string memory s) internal view returns (bytes memory) {
-    uint16[] memory _counters = new uint16[](2);
-    _counters[0] = uint16(a32.length * 4);
-    _counters[1] = uint16(bytes(s).length);
+    uint40[] memory _counters = new uint40[](2);
+    _counters[0] = uint40(a32.length * 4);
+    _counters[1] = uint40(bytes(s).length);
     PackedCounter _encodedLengths = PackedCounterLib.pack(_counters);
 
     return abi.encodePacked(u32, u128, _encodedLengths.unwrap(), EncodeArray.encode((a32)), bytes((s)));
