+++ packages/store/test/StoreCore.t.sol
@@ -303,9 +303,9 @@ contract StoreCoreTest is Test, StoreReadWithStubs {
 
     PackedCounter encodedDynamicLength;
     {
-      uint16[] memory dynamicLengths = new uint16[](2);
-      dynamicLengths[0] = uint16(secondDataBytes.length);
-      dynamicLengths[1] = uint16(thirdDataBytes.length);
+      uint40[] memory dynamicLengths = new uint40[](2);
+      dynamicLengths[0] = uint40(secondDataBytes.length);
+      dynamicLengths[1] = uint40(thirdDataBytes.length);
       encodedDynamicLength = PackedCounterLib.pack(dynamicLengths);
     }
 
@@ -487,7 +487,7 @@ contract StoreCoreTest is Test, StoreReadWithStubs {
     assertEq(keccak256(loadedData), keccak256(fourthDataBytes));
 
     // Verify all fields are correct
-    PackedCounter encodedLengths = PackedCounterLib.pack(uint16(thirdDataBytes.length), uint16(fourthDataBytes.length));
+    PackedCounter encodedLengths = PackedCounterLib.pack(uint40(thirdDataBytes.length), uint40(fourthDataBytes.length));
     assertEq(
       keccak256(StoreCore.getRecord(table, key)),
       keccak256(
@@ -524,9 +524,9 @@ contract StoreCoreTest is Test, StoreReadWithStubs {
 
     PackedCounter encodedDynamicLength;
     {
-      uint16[] memory dynamicLengths = new uint16[](2);
-      dynamicLengths[0] = uint16(secondDataBytes.length);
-      dynamicLengths[1] = uint16(thirdDataBytes.length);
+      uint40[] memory dynamicLengths = new uint40[](2);
+      dynamicLengths[0] = uint40(secondDataBytes.length);
+      dynamicLengths[1] = uint40(thirdDataBytes.length);
       encodedDynamicLength = PackedCounterLib.pack(dynamicLengths);
     }
 
@@ -777,9 +777,9 @@ contract StoreCoreTest is Test, StoreReadWithStubs {
 
     // startByteIndex must not overflow
     vm.expectRevert(
-      abi.encodeWithSelector(IStoreErrors.StoreCore_DataIndexOverflow.selector, type(uint16).max, type(uint32).max)
+      abi.encodeWithSelector(IStoreErrors.StoreCore_DataIndexOverflow.selector, type(uint40).max, type(uint56).max)
     );
-    StoreCore.updateInField(table, key, 2, type(uint32).max, thirdDataForUpdate);
+    StoreCore.updateInField(table, key, 2, type(uint56).max, thirdDataForUpdate);
   }
 
   function testAccessEmptyData() public {
@@ -872,7 +872,7 @@ contract StoreCoreTest is Test, StoreReadWithStubs {
     uint32[] memory arrayData = new uint32[](1);
     arrayData[0] = 0x01020304;
     bytes memory arrayDataBytes = EncodeArray.encode(arrayData);
-    PackedCounter encodedArrayDataLength = PackedCounterLib.pack(uint16(arrayDataBytes.length));
+    PackedCounter encodedArrayDataLength = PackedCounterLib.pack(uint40(arrayDataBytes.length));
     bytes memory dynamicData = abi.encodePacked(encodedArrayDataLength.unwrap(), arrayDataBytes);
     bytes memory staticData = abi.encodePacked(bytes16(0x0102030405060708090a0b0c0d0e0f10));
     bytes memory data = abi.encodePacked(staticData, dynamicData);
