+++ packages/store/ts/library/render-solidity/renderTable.ts
@@ -169,12 +169,12 @@ ${renderTypeHelpers(options)}
 function renderEncodedLengths(dynamicFields: RenderDynamicField[]) {
   if (dynamicFields.length > 0) {
     return `
-    uint16[] memory _counters = new uint16[](${dynamicFields.length});
+    uint40[] memory _counters = new uint40[](${dynamicFields.length});
     ${renderList(dynamicFields, ({ name, arrayElement }, index) => {
       if (arrayElement) {
-        return `_counters[${index}] = uint16(${name}.length * ${arrayElement.staticByteLength});`;
+        return `_counters[${index}] = uint40(${name}.length * ${arrayElement.staticByteLength});`;
       } else {
-        return `_counters[${index}] = uint16(bytes(${name}).length);`;
+        return `_counters[${index}] = uint40(bytes(${name}).length);`;
       }
     })}
     PackedCounter _encodedLengths = PackedCounterLib.pack(_counters);
